pyproject.toml
[project]
name = "cloud-to-local-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "huggingface-hub>=1.7.1",
    "langchain>=1.2.12",
    "langgraph>=1.1.2",
    "python-dotenv>=1.2.2",
    "transformers>=5.3.0",
]
